[ALPS05698443] DISP: 34-bit addrssing porting
34-bit addressing porting.
TO DO: need to check how to use secure cmdq to mask write
       for OVL addrss msb in SVP

MTK-Commit-Id: 3b57081202ecfefc9a159df3b8663bddb882865c

CR-Id: ALPS05698443
Feature: [Module]Display Driver

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_debug.c
@@ -1282,7 +1282,7 @@ static void mtk_drm_cwb_info_init(struct drm_crtc *crtc)
 
 		mtk_gem = mtk_drm_gem_create(
 			crtc->dev, mode.width * mode.height * 3, true);
-		cwb_info->buffer[0].addr_mva = (u32)mtk_gem->dma_addr;
+		cwb_info->buffer[0].addr_mva = mtk_gem->dma_addr;
 		cwb_info->buffer[0].addr_va = (u64)mtk_gem->kvaddr;
 
 		cwb_info->buffer[0].fb  =
@@ -1294,7 +1294,7 @@ static void mtk_drm_cwb_info_init(struct drm_crtc *crtc)
 
 		mtk_gem = mtk_drm_gem_create(
 			crtc->dev, mode.width * mode.height * 3, true);
-		cwb_info->buffer[1].addr_mva = (u32)mtk_gem->dma_addr;
+		cwb_info->buffer[1].addr_mva = mtk_gem->dma_addr;
 		cwb_info->buffer[1].addr_va = (u64)mtk_gem->kvaddr;
 
 		cwb_info->buffer[1].fb  =
@@ -1413,7 +1413,7 @@ bool mtk_drm_set_cwb_roi(struct mtk_rect rect)
 
 		mtk_gem = mtk_drm_gem_create(
 			crtc->dev, mode.width * mode.height * 3, true);
-		cwb_info->buffer[0].addr_mva = (u32)mtk_gem->dma_addr;
+		cwb_info->buffer[0].addr_mva = mtk_gem->dma_addr;
 		cwb_info->buffer[0].addr_va = (u64)mtk_gem->kvaddr;
 
 		cwb_info->buffer[0].fb  =
@@ -1425,7 +1425,7 @@ bool mtk_drm_set_cwb_roi(struct mtk_rect rect)
 
 		mtk_gem = mtk_drm_gem_create(
 			crtc->dev, mode.width * mode.height * 3, true);
-		cwb_info->buffer[1].addr_mva = (u32)mtk_gem->dma_addr;
+		cwb_info->buffer[1].addr_mva = mtk_gem->dma_addr;
 		cwb_info->buffer[1].addr_va = (u64)mtk_gem->kvaddr;
 
 		cwb_info->buffer[1].fb  =
